requirements.txt
absl-py==0.11.0
asgiref==3.2.10
astunparse==1.6.3
cachetools==4.2.0
certifi==2020.6.20
chardet==3.0.4
cmake==3.18.2.post1
cycler==0.10.0
Django==3.1
django-crispy-forms==1.9.2
dlib==19.21.0
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.34.0
h5py==2.10.0
idna==2.10
imutils==0.5.3
joblib==0.16.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.3.3
matplotlib==3.3.1
mysql==0.0.2
mysqlclient==2.0.1
numpy==1.18.5
oauthlib==3.1.0
opencv-contrib-python==4.4.0.42
opencv-python==4.4.0.42
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pyparsing==2.4.7
pytesseract==0.3.6
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.1
tensorflow-gpu-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
twilio==6.45.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
